requirements.txt
telethon 
tinydb
python-dotenv
python-telegram-bot